Remove useContractAbiItems
